requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
amqp==5.2.0
    # via
    #   -r requirements.in
    #   kombu
annotated-types==0.6.0
    # via pydantic
arabic-reshaper==3.0.0
    # via xhtml2pdf
asgiref==3.7.2
    # via
    #   django
    #   django-countries
asn1crypto==1.5.1
    # via
    #   oscrypto
    #   pyhanko
    #   pyhanko-certvalidator
async-timeout==4.0.3
    # via redis
attrs==21.2.0
    # via -r requirements.in
authlib==1.2.1
    # via -r requirements.in
billiard==4.2.0
    # via celery
brotli==1.1.0
    # via fonttools
celery==5.3.6
    # via -r requirements.in
certifi==2024.2.2
    # via requests
cffi==1.16.0
    # via
    #   cryptography
    #   weasyprint
chardet==5.2.0
    # via reportlab
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pyhanko
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
cryptography==42.0.5
    # via
    #   authlib
    #   pyhanko
    #   pyhanko-certvalidator
cssselect2==0.7.0
    # via
    #   svglib
    #   weasyprint
defusedxml==0.7.1
    # via odfpy
diff-match-patch==20230430
    # via django-import-export
django==5.0
    # via
    #   -r requirements.in
    #   django-auditlog
    #   django-betterforms
    #   django-bootstrap5
    #   django-dirtyfields
    #   django-filter
    #   django-formtools
    #   django-import-export
    #   django-jazzmin
    #   django-qr-code
    #   django-renderpdf
    #   django-tables2
    #   djangorestframework
django-active-link==0.1.8
    # via -r requirements.in
django-auditlog==2.3.0
    # via -r requirements.in
django-betterforms==2.0.0
    # via -r requirements.in
django-bootstrap5==23.3
    # via -r requirements.in
django-countries==7.5.1
    # via -r requirements.in
django-dirtyfields==1.9.2
    # via -r requirements.in
django-filter==23.5
    # via -r requirements.in
django-formtools==2.4.1
    # via -r requirements.in
django-import-export==3.3.4
    # via -r requirements.in
django-jazzmin==2.6.0
    # via -r requirements.in
django-qr-code==4.0.1
    # via -r requirements.in
django-renderpdf==4.0.0
    # via -r requirements.in
django-tables2==2.7.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
fontawesomefree==6.5.1
    # via -r requirements.in
fonttools[woff]==4.50.0
    # via weasyprint
freetype-py==2.3.0
    # via
    #   reportlab
    #   rlpycairo
html5lib==1.1
    # via
    #   weasyprint
    #   xhtml2pdf
idna==3.6
    # via requests
kombu==5.3.5
    # via celery
lxml==5.1.0
    # via svglib
markuppy==1.14
    # via tablib
mysqlclient==2.2.1
    # via -r requirements.in
odfpy==1.4.1
    # via tablib
openpyxl==3.1.2
    # via tablib
oscrypto==1.3.0
    # via pyhanko-certvalidator
pillow==10.2.0
    # via
    #   -r requirements.in
    #   reportlab
    #   weasyprint
    #   xhtml2pdf
prompt-toolkit==3.0.43
    # via click-repl
pycairo==1.26.0
    # via rlpycairo
pycparser==2.21
    # via cffi
pydantic==2.6.4
    # via django-qr-code
pydantic-core==2.16.3
    # via pydantic
pydyf==0.9.0
    # via weasyprint
pyhanko==0.23.1
    # via xhtml2pdf
pyhanko-certvalidator==0.26.3
    # via
    #   pyhanko
    #   xhtml2pdf
pypdf==4.1.0
    # via xhtml2pdf
pyphen==0.14.0
    # via weasyprint
pypng==0.20220715.0
    # via qrcode
python-bidi==0.4.2
    # via xhtml2pdf
python-dateutil==2.9.0.post0
    # via
    #   celery
    #   django-auditlog
python-magic==0.4.27
    # via -r requirements.in
python-slugify==8.0.1
    # via -r requirements.in
pytz==2024.1
    # via djangorestframework
pyyaml==6.0.1
    # via
    #   pyhanko
    #   tablib
qrcode==7.4.2
    # via pyhanko
redis==5.0.1
    # via -r requirements.in
reportlab[pycairo]==4.1.0
    # via
    #   svglib
    #   xhtml2pdf
requests==2.31.0
    # via
    #   -r requirements.in
    #   pyhanko
    #   pyhanko-certvalidator
rlpycairo==0.3.0
    # via reportlab
segno==1.6.1
    # via django-qr-code
six==1.16.0
    # via
    #   html5lib
    #   python-bidi
    #   python-dateutil
sqlparse==0.4.4
    # via django
svglib==1.5.1
    # via xhtml2pdf
tablib[html,ods,xls,xlsx,yaml]==3.5.0
    # via
    #   -r requirements.in
    #   django-import-export
text-unidecode==1.3
    # via python-slugify
tinycss2==1.2.1
    # via
    #   cssselect2
    #   svglib
    #   weasyprint
typing-extensions==4.10.0
    # via
    #   asgiref
    #   django-countries
    #   pydantic
    #   pydantic-core
    #   qrcode
tzdata==2024.1
    # via celery
tzlocal==5.2
    # via pyhanko
uritools==4.0.2
    # via pyhanko-certvalidator
urllib3==2.2.1
    # via requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
weasyprint==60.2
    # via
    #   -r requirements.in
    #   django-renderpdf
webencodings==0.5.1
    # via
    #   cssselect2
    #   html5lib
    #   tinycss2
xhtml2pdf==0.2.13
    # via -r requirements.in
xlrd==2.0.1
    # via tablib
xlwt==1.3.0
    # via tablib
zopfli==0.2.3
    # via fonttools
